+++ src/main/java/org/prgms/voucherProgram/domain/customer/exception/CustomerIsNotExistsException.java
@@ -1,6 +1,8 @@
 package org.prgms.voucherProgram.domain.customer.exception;
 
-public class CustomerIsNotExistsException extends IllegalArgumentException {
+import org.prgms.voucherProgram.global.error.exception.EntityNotFoundException;
+
+public class CustomerIsNotExistsException extends EntityNotFoundException {
     private static final String MESSAGE = "[ERROR] 해당 이메일로 저장된 고객이 없습니다.";
 
     public CustomerIsNotExistsException() {
