+++ src/main/java/org/prgms/voucherProgram/domain/voucher/exception/VoucherIsNotExistsException.java
@@ -1,6 +1,8 @@
 package org.prgms.voucherProgram.domain.voucher.exception;
 
-public class VoucherIsNotExistsException extends IllegalArgumentException {
+import org.prgms.voucherProgram.global.error.exception.EntityNotFoundException;
+
+public class VoucherIsNotExistsException extends EntityNotFoundException {
     private static final String MESSAGE = "[ERROR] 해당 아이디로 저장된 바우처가 없습니다.";
 
     public VoucherIsNotExistsException() {
