+++ src/main/java/org/prgms/voucherProgram/global/error/ExceptionControllerAdvice.java
@@ -1,5 +1,6 @@
-package org.prgms.voucherProgram.global.exception;
+package org.prgms.voucherProgram.global.error;
 
+import org.prgms.voucherProgram.global.error.exception.EntityNotFoundException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.ui.Model;
@@ -11,13 +12,19 @@ public class ExceptionControllerAdvice {
 
     private static final Logger logger = LoggerFactory.getLogger(ExceptionControllerAdvice.class);
 
-    @ExceptionHandler
+    @ExceptionHandler({EntityNotFoundException.class})
+    public String handleEntityNotFoundException(EntityNotFoundException e, Model model) {
+        model.addAttribute("error", e.getMessage());
+        return "error";
+    }
+
+    @ExceptionHandler(IllegalArgumentException.class)
     public String handleCustomerException(IllegalArgumentException e, Model model) {
         model.addAttribute("error", e.getMessage());
         return "error";
     }
 
-    @ExceptionHandler
+    @ExceptionHandler(Exception.class)
     public String handleAllException(Exception e, Model model) {
         logger.error(e.getMessage());
         model.addAttribute("error", e.getMessage());
