+++ src/main/java/org/prgms/voucherProgram/global/error/ExceptionResponse.java
@@ -1,4 +1,4 @@
-package org.prgms.voucherProgram.global.exception;
+package org.prgms.voucherProgram.global.error;
 
 import java.util.Date;
