+++ src/main/java/org/prgms/voucherProgram/global/error/exception/NothingChangeException.java
@@ -1,4 +1,4 @@
-package org.prgms.voucherProgram.global.exception;
+package org.prgms.voucherProgram.global.error.exception;
 
 public class NothingChangeException extends RuntimeException {
     private static final String MESSAGE = "[ERROR] 해당 요청이 정상적으로 처리되지 않았습니다.";
