+++ src/test/java/org/prgms/voucherProgram/domain/customer/repository/JdbcCustomerRepositoryTest.java
@@ -25,7 +25,7 @@ import org.prgms.voucherProgram.domain.voucher.domain.FixedAmountVoucher;
 import org.prgms.voucherProgram.domain.voucher.domain.Voucher;
 import org.prgms.voucherProgram.domain.voucher.repository.JdbcVoucherRepository;
 import org.prgms.voucherProgram.domain.voucher.repository.VoucherRepository;
-import org.prgms.voucherProgram.global.exception.NothingChangeException;
+import org.prgms.voucherProgram.global.error.exception.NothingChangeException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.jdbc.DataSourceBuilder;
 import org.springframework.context.annotation.Bean;
